TINY-8729: move deleteRangeContents to more shared DeleteUtils
