+++ modules/tinymce/src/core/main/ts/delete/TableDelete.ts
@@ -1,5 +1,5 @@
 import { Arr, Fun, Optional, Optionals } from '@ephox/katamari';
-import { Attribute, Compare, Remove, SugarElement, SugarNode, Traverse } from '@ephox/sugar';
+import { Attribute, Compare, Remove, SugarElement, SugarNode } from '@ephox/sugar';
 
 import Editor from '../api/Editor';
 import * as CaretFinder from '../caret/CaretFinder';
@@ -12,14 +12,10 @@ import * as PaddingBr from '../dom/PaddingBr';
 import * as Parents from '../dom/Parents';
 import * as TableCellSelection from '../selection/TableCellSelection';
 import * as DeleteElement from './DeleteElement';
+import { deleteRangeContents } from './DeleteUtils';
 import * as TableDeleteAction from './TableDeleteAction';
 
 type OutsideTableDetails = TableDeleteAction.OutsideTableDetails;
-
-const freefallRtl = (root: SugarElement<Node>): Optional<SugarElement<Node>> => {
-  const child = SugarNode.isComment(root) ? Traverse.prevSibling(root) : Traverse.lastChild(root);
-  return child.bind(freefallRtl).orThunk(() => Optional.some(root));
-};
 
 // Reset the contenteditable state and fill the content with a padding br
 const cleanCells = (cells: SugarElement<HTMLTableCellElement>[]): void =>
@@ -150,25 +146,7 @@ const emptyMultiTableCells = (
 // Delete the contents of a range inside a cell. Runs on tables that are a single cell or partial selections that need to be cleaned up.
 const deleteCellContents = (editor: Editor, rng: Range, cell: SugarElement<HTMLTableCellElement>, moveSelection: boolean = true): Optional<() => void> =>
   Optional.some(() => {
-    rng.deleteContents();
-    // Pad the last block node
-    const lastNode = freefallRtl(cell).getOr(cell);
-    const lastBlock = SugarElement.fromDom(editor.dom.getParent(lastNode.dom, editor.dom.isBlock));
-    if (Empty.isEmpty(lastBlock)) {
-      PaddingBr.fillWithPaddingBr(lastBlock);
-      if (moveSelection) {
-        editor.selection.setCursorLocation(lastBlock.dom, 0);
-      }
-    }
-    // Clean up any additional leftover nodes. If the last block wasn't a direct child, then we also need to clean up siblings
-    if (!Compare.eq(cell, lastBlock)) {
-      const additionalCleanupNodes = Optionals.is(Traverse.parent(lastBlock), cell) ? [] : Traverse.siblings(lastBlock);
-      Arr.each(additionalCleanupNodes.concat(Traverse.children(cell)), (node) => {
-        if (!Compare.eq(node, lastBlock) && !Compare.contains(node, lastBlock) && Empty.isEmpty(node)) {
-          Remove.remove(node);
-        }
-      });
-    }
+    deleteRangeContents(editor, rng, cell, moveSelection);
   });
 
 const deleteTableElement = (editor: Editor, table: SugarElement<HTMLTableElement>): Optional<() => void> =>
